Fixed signed/unsigned comparisons. Should be perhaps casting to ssize_t instead of int.
[git-p4: depot-paths = "//depot/openwatcom/": change = 29239]

+++ bld/clib/string/c/rstrncpy.c
@@ -24,8 +24,7 @@
 *
 *  ========================================================================
 *
-* Description:  WHEN YOU FIGURE OUT WHAT THIS FILE DOES, PLEASE
-*               DESCRIBE IT HERE!
+* Description:  RISC style strncpy().
 *
 ****************************************************************************/
 
@@ -52,7 +51,7 @@ CHAR_TYPE *__F_NAME(strncpy,wcsncpy)( CHAR_TYPE *dest, const CHAR_TYPE *src,
         }
     #endif
 
-    if( n <= 0 )  return( destStart );
+    if( (int)n <= 0 )  return( destStart );
 
     /*** Copy any unaligned bytes at the start (align dest pointer) ***/
     while( offset != 0  &&  n >= 1 ) {
